Fix contributor stats endpoint routing
