+++ nwleaderboard-api/src/main/java/com/opyruso/nwleaderboard/ContributorResource.java
@@ -5,7 +5,6 @@ import com.opyruso.nwleaderboard.dto.ContributionExtractionResponseDto;
 import com.opyruso.nwleaderboard.dto.ContributionRunDto;
 import com.opyruso.nwleaderboard.dto.ContributionScanDetailDto;
 import com.opyruso.nwleaderboard.dto.ContributionScanSummaryDto;
-import com.opyruso.nwleaderboard.dto.ContributorWeeklyRunsResponse;
 import com.opyruso.nwleaderboard.dto.UpdateContributionScanRequest;
 import com.opyruso.nwleaderboard.dto.UpdateDungeonHighlightsRequest;
 import com.opyruso.nwleaderboard.dto.RegionResponse;
@@ -14,7 +13,6 @@ import com.opyruso.nwleaderboard.service.ContributorExtractionService;
 import com.opyruso.nwleaderboard.service.ContributorExtractionService.ContributorRequestException;
 import com.opyruso.nwleaderboard.service.ContributorSubmissionService;
 import com.opyruso.nwleaderboard.service.ContributorSubmissionService.ContributorSubmissionException;
-import com.opyruso.nwleaderboard.service.ContributorStatisticsService;
 import com.opyruso.nwleaderboard.service.DungeonService;
 import com.opyruso.nwleaderboard.service.RegionService;
 import com.opyruso.nwleaderboard.service.ScanLeaderboardService;
@@ -65,10 +63,7 @@ public class ContributorResource {
 
     @Inject
     DungeonService dungeonService;
-
-    @Inject
-    ContributorStatisticsService statisticsService;
-
+    
     @Inject
     ScanLeaderboardService scanLeaderboardService;
 
@@ -146,25 +141,6 @@ public class ContributorResource {
             LOG.error("Unable to update highlighted dungeons", e);
             return Response.status(Status.BAD_GATEWAY)
                     .entity(new ApiMessageResponse("Unable to update dungeon highlights", null))
-                    .build();
-        }
-    }
-
-    @GET
-    @Path("/runs/weekly")
-    public Response listRunsByWeek() {
-        if (!hasContributorRole()) {
-            return Response.status(Status.FORBIDDEN)
-                    .entity(new ApiMessageResponse("Contributor role required", null))
-                    .build();
-        }
-        try {
-            List<ContributorWeeklyRunsResponse> summaries = statisticsService.listRunsByWeek();
-            return Response.ok(summaries).build();
-        } catch (Exception e) {
-            LOG.error("Unable to load contributor run statistics", e);
-            return Response.status(Status.BAD_GATEWAY)
-                    .entity(new ApiMessageResponse("Unable to load run statistics", null))
                     .build();
         }
     }
